requirements.txt
websockets~=11.0
pydantic>=2.5
python-dotenv>=1.0